+++ CHANGELOG.md
@@ -4,8 +4,9 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
-- You can now use `zacs:style` on all elements, not just ZACS-styled elements.
-- `'.foo, .bar': {}` CSS-only stylesets are now allowed
+- NEW: You can now use `zacs:style` on all elements
+- NEW: `'.foo, .bar': {}` CSS-only stylesets
+- NEW: `border: [1, 'red']` shorthand
 - `__zacs_original_name` debug attribute is now placed at the beginning of output elements
 - `className={}` is now consistently placed before `style={}`
 - stylesheets: overriden properties are stripped on native to save space and silence Hermes warnings